Instead of creating a fake test command, try to iterate over existing commands structure and replace only what's needed
